requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.3.1
ecdsa==0.19.2
fastapi==0.135.3
Flask==3.1.3
flexcache==0.3
flexparser==0.4
fonttools==4.62.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
limits==5.8.0
MarkupSafe==3.0.3
matplotlib==3.10.8
MIDIUtil==1.2.1
mido==1.3.3
MouseInfo==0.1.3
numpy==2.4.4
obd==0.7.3
packaging==26.0
pandas==3.0.2
passlib==1.7.4
pillow==12.2.0
Pint==0.24.4
platformdirs==4.9.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.27
python-rtmidi==1.5.8
pytweening==1.2.0
requests==2.33.1
rsa==4.9.1
rubicon-objc==0.5.3
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
slowapi==0.1.9
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
Werkzeug==3.1.6
wrapt==2.1.2
